+++ lib/src/cupertino/cupertino.dart
@@ -1,1 +1,2 @@
 export 'containers/containers.dart';
+export 'control/control.dart';